feat(phase-4): NET and STR HTTP handlers
- handlers_net.go: GET /v1/net/summary|orchestrators|orchestrators/{addr}|gpu|models
- handlers_stream.go: GET /v1/streams/active|summary|history
- writeError helper: RFC 7807 problem+json error responses
- Wire 8 real routes in buildRouter; remaining Phase 5 routes stay as 501 stubs
- Fix UInt64→int64 scan type mismatch in all ClickHouse repo methods
  (count()/sum() return UInt64; scan into uint64 intermediates then cast)
- handlers_net_test.go + handlers_stream_test.go: happy-path, param, and
  404/silent-fallback tests using NoopAnalyticsRepo
- Validated against live ClickHouse: 85 registered orchs, real data returned

+++ api/internal/repo/clickhouse/network.go
@@ -26,7 +26,8 @@ func (r *Repo) GetNetworkSummary(ctx context.Context, p types.QueryParams) (*typ
 	}
 
 	row := r.conn.QueryRow(ctx, query, args...)
-	var total, active int64
+	// count() returns UInt64 in ClickHouse; scan into uint64 then cast.
+	var total, active uint64
 	if err := row.Scan(&total, &active); err != nil {
 		return nil, fmt.Errorf("clickhouse get network summary: %w", err)
 	}
@@ -34,8 +35,8 @@ func (r *Repo) GetNetworkSummary(ctx context.Context, p types.QueryParams) (*typ
 	return &types.NetworkSummary{
 		Org:                    p.Org,
 		SnapshotTime:           time.Now().UTC(),
-		TotalRegistered:        total,
-		TotalActive:            active,
+		TotalRegistered:        int64(total),
+		TotalActive:            int64(active),
 		ActiveThresholdMinutes: activeOrchMinutes,
 	}, nil
 }
